rename utils.js config.js
